start to implement propagation
It looks like the propagator can really benefit from propagation beyond
conflict detection. This can be seen with the simple example and the
current prototypical implemention, which finds all consequences given
the current state of the tableaux by fully traversing it. It is possible
to implement this more efficiently by mainting counters that are updated
in response to changes of the tableaux.